updated VS
