+++ GigaChat.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30711.63
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34616.47
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GigaChat", "WpfApp1\GigaChat.csproj", "{DECE010E-501C-407C-9A5D-255534875114}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GigaChat", "GigaChat\GigaChat.csproj", "{4F429110-7354-40E0-9C1A-219BC9529700}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DECE010E-501C-407C-9A5D-255534875114}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DECE010E-501C-407C-9A5D-255534875114}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DECE010E-501C-407C-9A5D-255534875114}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DECE010E-501C-407C-9A5D-255534875114}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4F429110-7354-40E0-9C1A-219BC9529700}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F429110-7354-40E0-9C1A-219BC9529700}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F429110-7354-40E0-9C1A-219BC9529700}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F429110-7354-40E0-9C1A-219BC9529700}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {EFD6F245-8FD2-4EC0-8E6E-06C095150D0F}
+		SolutionGuid = {6AE1B049-531A-4D5D-9C74-65146CB50C46}
 	EndGlobalSection
 EndGlobal
