+++ GigaChat/App.xaml
@@ -2,8 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:GigaChat"
-             StartupUri="LoaderWindow.xaml">
+             StartupUri="MainWindow.xaml">
     <Application.Resources>
-        
+         
     </Application.Resources>
 </Application>
