+++ GigaChat/LoaderWindow.xaml.cs
