+++ GigaChat/MainWindow.xaml
@@ -5,7 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:GigaChat"
         mc:Ignorable="d"
-        Title="GigaChat" Height="600" Width="900" Background="#FF1B1B1B" SizeChanged="OnWindowResize" MinWidth="900" MinHeight="600" Initialized="mainWindow_Initialized" ScrollViewer.VerticalScrollBarVisibility="Disabled" WindowStyle="None">
+        Title="GigaChat" Height="600" Width="900" Background="#FF1B1B1B" SizeChanged="OnWindowResize" MinWidth="900" MinHeight="600" Initialized="mainWindow_Initialized" ScrollViewer.VerticalScrollBarVisibility="Disabled" WindowStyle="None" Icon="/baseLogo.jpg">
     <StackPanel x:Name="WindowLayout" Background="Black" Margin="0,0,0,0">
         <StackPanel x:Name="SystemWindowPanel" Height="26" RenderTransformOrigin="0.5,0.5" Background="#FF323232" Margin="0,0,0,0" VerticalAlignment="Center" Orientation="Horizontal">
             <StackPanel.RenderTransform>
@@ -18,31 +18,15 @@
             </StackPanel.RenderTransform>
             <Label x:Name="GigaChatLabel" Content="GiGaChat" HorizontalAlignment="Left" HorizontalContentAlignment="Center" VerticalContentAlignment="Stretch" FontFamily="Magneto" FontSize="16" Background="{x:Null}" Foreground="White"/>
             <StackPanel x:Name="SystemTopPanel" Width="720"/>
-            <Image x:Name="HideButton" Width="13" Height="13" MinWidth="13" MinHeight="13" MaxWidth="26" MaxHeight="26" Source="windowHide.png" MouseLeftButtonDown="HideButton_MouseLeftButtonDown">
-                <Image.OpacityMask>
-                    <ImageBrush ImageSource="windowHide.png"/>
-                </Image.OpacityMask>
-            </Image>
-            <Image x:Name="ModeButton" Width="13" MinWidth="13" MinHeight="13" Height="13" MaxWidth="13" MaxHeight="13" Source="windowHigh.png" Margin="10,0,0,0" MouseDown="ModeButton_MouseDown">
-                <Image.OpacityMask>
-                    <ImageBrush ImageSource="windowLow.png"/>
-                </Image.OpacityMask>
-            </Image>
-            <Image x:Name="ExitButton" Width="13" Source="exit.png" Height="13" MaxWidth="13" MaxHeight="13" MinWidth="13" MinHeight="13" Margin="10,0,0,0" HorizontalAlignment="Right" MouseLeftButtonDown="ExitButton_MouseLeftButtonDown">
-                <Image.OpacityMask>
-                    <ImageBrush ImageSource="exit.png"/>
-                </Image.OpacityMask>
-            </Image>
+            <Image x:Name="HideButton" Width="13" Height="13" MinWidth="13" MinHeight="13" MaxWidth="26" MaxHeight="26" Source="/windowHide.png" MouseLeftButtonDown="HideButton_MouseLeftButtonDown"/>
+            <Image x:Name="ModeButton" Width="13" MinWidth="13" MinHeight="13" Height="13" MaxWidth="13" MaxHeight="13" Source="/windowHigh.png" Margin="10,0,0,0" MouseDown="ModeButton_MouseDown"/>
+            <Image x:Name="ExitButton" Width="13" Source="/exit.png" Height="13" MaxWidth="13" MaxHeight="13" MinWidth="13" MinHeight="13" Margin="10,0,0,0" HorizontalAlignment="Right" MouseLeftButtonDown="ExitButton_MouseLeftButtonDown"/>
         </StackPanel>
         <StackPanel x:Name="ControlDataPanel" Height="564" Background="#FF1B1B1B" Orientation="Horizontal">
             <StackPanel x:Name="ControlPanel" Width="90" Background="#FF191919" Height="573" VerticalAlignment="Bottom" Margin="0,0,0,-9">
-                <Image x:Name="SettingsButton" Height="70" Source="settings.png" Width="70" Margin="0">
-                    <Image.OpacityMask>
-                        <ImageBrush ImageSource="settings.png"/>
-                    </Image.OpacityMask>
-                </Image>
+                <Image x:Name="SettingsButton" Height="70" Source="/settings.png" Width="70" Margin="0"/>
                 <Rectangle x:Name="ChannelsLine" Fill="#FFF4F4F5" Height="3"  Margin="8,0" RadiusY="1.375" RadiusX="1.375"/>
-                <ScrollViewer x:Name="ChannelPanelScrollViewer" Width="90" HorizontalAlignment="Left" Cursor="Hand">
+                <ScrollViewer x:Name="ChannelPanelScrollViewer" Width="90" HorizontalAlignment="Left" Cursor="Hand" Visibility="Hidden">
                     <StackPanel x:Name="ChannelsPanel" Height="495" CanVerticallyScroll="True" ScrollViewer.CanContentScroll="True">
                         <StackPanel.Background>
                             <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
